move to existing utils.ts
